pyproject.toml
[tool.poetry]
name = "HttpTesting"
version = "1.1.97"
description = "Httptesting HTTP(s) interface testing framework."
license = "Apache-2.0"
readme = "README.md"
authors = ["天枢 <lengyaohui@163.com>"]

homepage = "https://github.com/HttpTesting/pyhttp"
repository = "https://github.com/HttpTesting/pyhttp"


keywords = ["http", "api", "test", "requests"]
classifiers  =  [ 
    "Environment :: Console",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Operating System :: Microsoft :: Windows",
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
]

include = ["LICENSE", "httptesting/templates/*", "httptesting/config/*", "httptesting/case/*"]

[tool.poetry.dependencies]
Python = "^3.5"
pyyaml = "5.1.1"
pytest = "5.0.1"
pytest-html = "1.21.1"
pytest-repeat = "0.8.0"
pytest-rerunfailures = "7.0"
pytest-xdist = "1.29.0"
requests = "2.22.0"
requests-toolbelt = "0.9.1"
colorama = "0.4.1"
[tool.poetry.dev-dependencies]


[tool.poetry.scripts]
am = "httptesting.main:run_min"
amt = "httptesting.main:run_min"


[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"